requirements.txt
appium-python-client
selenium
pytest
pytest-xdist
allure-pytest